refactor(core): Extract JSON Schema helpers into a utils file
Move fixSchema and lockAdditionalProperties (+ its private helpers) out of runtime/tool-adapter.ts into utils/json-schema.ts — pure schema-in/schema-out transforms with no AI SDK dependency. tool-adapter and agent-runtime now import from there; the unit tests move to utils/__tests__/json-schema.test.ts (with added direct fixSchema coverage).
